third_party/WebKit/LayoutTests/inspector/console/console-object-preview-accessors-expected.txt

CONSOLE MESSAGE: line 2: [object Object],[object Object]
CONSOLE MESSAGE: line 2: [object Object]
Tests that console produces instant previews for arrays and objects that have getters and setters.

Before expanding accessors:
arrayLikeWithAccessors
objectWithAccessors
console.table(tableObject)
console.dir(objectWithAccessors)
[(...), (...), undefined × 3, (...), 6, (...), "value8", (...)]
Object {0: (...), 1: (...), 3: (...), 5: (...), 6: 6, 7: (...), 8: "value8", 9: (...)}
HEADER (index) | 0 | 1 | 3 | 5 | 6 | 7 | 8 | 9 | splice | 
ROW (...) | (...) | (...) | (...) | 6 | (...) | "value8" | (...) | function | 
ROW (...) | (...) | (...) | (...) | 6 | (...) | "value8" | (...) | 
undefined
Object0: (...)get 0: function 0() { baz.baz = 3; }1: (...)get 1: function 1() { throw new Error("objectWithAccessors"); }set 3: function 3() {}5: (...)get 5: function 5() { return "{property5}"; }6: 67: (...)get 7: function 7() { return 7788; }8: "value8"9: (...)get 9: function 9() { return -Infinity; }__proto__: Object VM:2
undefined

Accessor property buttons count: 28

After expanding accessors:
arrayLikeWithAccessors
objectWithAccessors
console.table(tableObject)
console.dir(objectWithAccessors)
[<exception>, <exception>, undefined × 3, "[property5]", 6, 77, "value8", NaN]
Object {0: <exception>, 1: <exception>, 3: undefined, 5: "{property5}", 6: 6, 7: 7788, 8: "value8", 9: -Infinity}
HEADER (index) | 0 | 1 | 3 | 5 | 6 | 7 | 8 | 9 | splice | 
ROW <exception> | <exception> | undefined | "[property5]" | 6 | 77 | "value8" | NaN | function | 
ROW <exception> | <exception> | undefined | "{property5}" | 6 | 7788 | "value8" | -Infinity | 
undefined
Object0: [Exception: ReferenceError: baz is not defined]get 0: function 0() { baz.baz = 3; }1: [Exception: Error: objectWithAccessors]get 1: function 1() { throw new Error("objectWithAccessors"); }set 3: function 3() {}5: "{property5}"get 5: function 5() { return "{property5}"; }6: 67: 7788get 7: function 7() { return 7788; }8: "value8"9: -Infinityget 9: function 9() { return -Infinity; }__proto__: Object VM:2
undefined
